refactor(runtime): 统一轮次选项与上下文上限建模
- 引入 PreparedTurnOptions 并收敛 provider 轮次准备入口

- ContextTokenLimit 与 models 配置校验改为 usize 宽度

- Responses 流输出项类型化并补充反序列化测试

- 抽取 requested_order_sort_key 并在 persistence/session-store 复用排序

- 简化 terminal-app prompt assembly 命令的错误传播

+++ crates/terminal-ui/src/context_budget/heatmap.rs
@@ -264,7 +264,7 @@ mod tests {
         }
     }
 
-    fn limit(value: u32) -> ContextTokenLimit {
+    fn limit(value: usize) -> ContextTokenLimit {
         ContextTokenLimit::try_from(value).expect("fixture limit should be valid")
     }
